+++ src/main/java/hu/oe/nik/szfmv/automatedcar/model/objects/CrossWalk.java
@@ -1,6 +1,6 @@
-package hu.oe.nik.szfmv.automatedcar.model.Objects;
+package hu.oe.nik.szfmv.automatedcar.model.objects;
 
-public abstract class Stationary extends Collidable {
+public class CrossWalk extends Crossable {
     /**
      * Creates an object of the virtual world on the given coordinates with the given image.
      *
@@ -8,15 +8,14 @@ public abstract class Stationary extends Collidable {
      * @param y             the initial y coordinate of the object
      * @param imageFileName the filename of the image representing the object in the virtual world
      */
-    public Stationary(int x, int y, String imageFileName) {
-
+    public CrossWalk(int x, int y, String imageFileName) {
         super(x, y, imageFileName);
     }
 
     /**
      * Creates an object with default parameter values.
      */
-    public Stationary() {
+    public CrossWalk() {
         super(0, 0, null);
     }
 }
