+++ src/main/java/hu/oe/nik/szfmv/automatedcar/model/objects/Stationary.java
@@ -1,6 +1,6 @@
-package hu.oe.nik.szfmv.automatedcar.model.Objects;
+package hu.oe.nik.szfmv.automatedcar.model.objects;
 
-public class ParkingPlace extends Crossable {
+public abstract class Stationary extends Collidable {
     /**
      * Creates an object of the virtual world on the given coordinates with the given image.
      *
@@ -8,14 +8,15 @@ public class ParkingPlace extends Crossable {
      * @param y             the initial y coordinate of the object
      * @param imageFileName the filename of the image representing the object in the virtual world
      */
-    public ParkingPlace(int x, int y, String imageFileName) {
+    public Stationary(int x, int y, String imageFileName) {
+
         super(x, y, imageFileName);
     }
 
     /**
      * Creates an object with default parameter values.
      */
-    public ParkingPlace() {
+    public Stationary() {
         super(0, 0, null);
     }
 }
